Extract app interface
