+++ internal/lintapp/lintapp.go
@@ -1,4 +1,4 @@
-package app
+package lintapp
 
 import (
 	"context"
@@ -6,35 +6,38 @@ import (
 	"os"
 	"os/exec"
 
+	"github.com/anttiharju/vmatch/pkg/app"
 	"github.com/anttiharju/vmatch/pkg/exit"
 	"github.com/anttiharju/vmatch/pkg/exit/exitcode"
 	"github.com/anttiharju/vmatch/pkg/pathfinder"
 	"github.com/anttiharju/vmatch/pkg/versionfinder"
 )
 
-type App struct {
+type LintApp struct {
 	desiredVersion string
 	installPath    string
 }
 
-func NewApp(versionFileName string) *App {
+var _ app.Interface = (*LintApp)(nil)
+
+func NewApp(versionFileName string) *LintApp {
 	desiredVersion := versionfinder.GetVersion(pathfinder.GetWorkDir(), versionFileName)
 	installPath := pathfinder.GetInstallPath(desiredVersion)
 
-	return &App{
+	return &LintApp{
 		desiredVersion: desiredVersion,
 		installPath:    installPath,
 	}
 }
 
-func (a *App) Run(ctx context.Context) int {
-	if a.noBinary() {
-		a.install(ctx)
+func (l *LintApp) Run(ctx context.Context) int {
+	if l.noBinary() {
+		l.install(ctx)
 	}
 
 	args := os.Args[1:]
 	//nolint:gosec // I don't think a wrapper can avoid G204.
-	linter := exec.Command(a.getGolangCILintPath(), args...)
+	linter := exec.Command(l.getGolangCILintPath(), args...)
 	linterOutput, _ := linter.Output()
 
 	fmt.Print(string(linterOutput))
@@ -42,19 +45,19 @@ func (a *App) Run(ctx context.Context) int {
 	return linter.ProcessState.ExitCode()
 }
 
-func (a *App) noBinary() bool {
-	_, err := os.Stat(a.getGolangCILintPath())
+func (l *LintApp) noBinary() bool {
+	_, err := os.Stat(l.getGolangCILintPath())
 
 	return os.IsNotExist(err)
 }
 
-func (a *App) install(_ context.Context) {
+func (l *LintApp) install(_ context.Context) {
 	//nolint:lll // Official binary install command:
 	// curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.59.1
 	curl := "curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh"
 	pipe := " | "
 	sh := "sh -s -- -b "
-	command := curl + pipe + sh + a.installPath + " v" + a.desiredVersion
+	command := curl + pipe + sh + l.installPath + " v" + l.desiredVersion
 	cmd := exec.Command("sh", "-c", command)
 
 	err := cmd.Start()
@@ -68,6 +71,6 @@ func (a *App) install(_ context.Context) {
 	}
 }
 
-func (a *App) getGolangCILintPath() string {
-	return a.installPath + string(os.PathSeparator) + "golangci-lint"
+func (l *LintApp) getGolangCILintPath() string {
+	return l.installPath + string(os.PathSeparator) + "golangci-lint"
 }
